pyproject.toml
[project]
name = "mcp-a2a-upskilling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.4",
    "acp-sdk>=1.0.3",
    "crewai>=0.175.0",
    "crewai-tools>=0.65.0",
    "ddgs>=9.5.5",
    "duckduckgo-search>=8.1.1",
    "flask>=3.1.2",
    "markdownify>=1.2.0",
    "mcp>=1.13.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.104.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "smolagents[mcp]>=1.21.3",
]
